users_interact_api/src/requirements.txt
aiokafka==0.8.1
aiohttp==3.8.5
pydantic[email]==1.9.1
uvicorn==0.23.2
python-dotenv==1.0
fastapi==0.100.1
orjson==3.9.2
gunicorn==21.2.0
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
httptools==0.6.0
motor==3.2.0
sentry-sdk==1.29.2
python-logstash==0.4.8